feat(exit): room requirements now clearly presented to the player
This change adds a TextMesh Pro mesh above the exit which reads
"Exit" until stepped over.

If the player is missing a puzzle piece, the message changes
informing them of what they need. It changes back when
they leave the trigger area.

If the player enters the trigger area with the puzzle piece
however, they are congratulated and a game reset is invoked
after a short pause.

Additionally, the entrance and exit are split out from the room
prefab and into their own so that they can be reused.