docs(readme): fix link and add notes to resolve a broken player armature
When cloning the repo and importing the project, the player materials
break.

To fix them, each of the materials on Player/Armature_Mesh need to
be set to Standard.

Ideally, it would be nice to have all dependencies automatically
download at project import time, but I've yet to figure out how
to do that for Asset Store packages.