fix(input): remap fire from West and East to Right Trigger
With the new camera system in place, it became clear that firing
the projectile with a button whilst also using the two sticks is
near impossible. Moving it to the right trigger feels better.

This commit makes the adjustment and also updates the controller
layout graphics in the tutorial and pause screens.