fix(player): numerous character and control-related isses resolved
This commit fixes regressions in the projectile since switching to the
new camera system. It was colliding wih the player and resolved by
moving it slightly ahead and turning on Shape -> Cone with appropriate
settings.

It also introduces a physic material to the moving platforms along
with project-wide physics settings that allow for smoother motion
when riding moving platforms.

Additionally, the capsule collider has been removed from the player
so as to not cause particular events to fire twice.