feat(journal): add core journal functionality
This commit adds a story journal which can be opened with either J
key or North on a gamepad. It can be navigated using left/right arrow
keys or left/right on a d-pad.

So far, it is capable of showing placeholder text from scriptable objects,
when we have permission to do so. For the purpose of testing, all
story cards are unlocked in the Unity editor for the time being.